[Refactor] 불필요한 주석, 코드, 스펠링 수정
