--- ARCollab/Assets/DebugLog.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6abae07dfb9005846bdadc34cf810a72
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
