+++ ARCollab/Assets/StreamingAssets/HiddenARCore.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 94f4a7c614338b148b51a0235b610100
-PrefabImporter:
+guid: f73fac98d807bf940823e60f955259bd
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
