+++ ARCollab/ProjectSettings/ProjectSettings.asset
@@ -136,6 +136,8 @@ PlayerSettings:
     Others: 1
   bundleVersion: 0.1
   preloadedAssets:
+  - {fileID: 4800000, guid: c9f956787b1d945e7b36e0516201fc76, type: 3}
+  - {fileID: 1227923878938332676, guid: 016c560e252006b42919b7307e5475c0, type: 2}
   - {fileID: 11400000, guid: 36f4f2c708645dd4cba7ac4cbde8529f, type: 2}
   metroInputSource: 0
   wsaTransparentSwapchain: 0
@@ -162,8 +164,8 @@ PlayerSettings:
     tvOS: 0
   overrideDefaultApplicationIdentifier: 0
   AndroidBundleVersionCode: 1
-  AndroidMinSdkVersion: 29
-  AndroidTargetSdkVersion: 29
+  AndroidMinSdkVersion: 24
+  AndroidTargetSdkVersion: 0
   AndroidPreferredInstallLocation: 1
   aotOptions: 
   stripEngineCode: 1
